Add loading screen to dashboard UI (#144)
This merge adds a loading screen to the dashboard UI that appears on initial UI page loads. The screen waits for the server to confirm that its connection to K8S is ready. This is helpful for remote servers where connection initialization times can take a few seconds. Eventually this should get replaced with more fine-grained connection status updates.

More detailed changes:
* Upgrades pnpm => 9.14.2
* Uses <Suspense /> and Apollo useSuspenseQuery to implement loading screen
* Adds use of <ErrorBoundary /> from 'react-error-boundary';