🐋 Modify cluster-agent docker files so that `rgkl` can be called without linker (#283)
## Summary

This PR modifies the production and development cluster-agent docker
files such that we can build a glibc dynamically-linked `rgkl` binary
that can be run from go code without the explicit use of a linker. This
simplifies the docker files themselves and also makes it easier to use
statically-linked `rgkl` binaries in development.

## Changes

Modify cluster-agent production and development docker files so that
required libraries to run rgkl are copied into the final image in their
expected places so that `rgkl` can be run from go code without the
explicit use of a linker.

## Submitter checklist

- [x] Add the correct emoji to PR title
- [x] Link the issue number to *Fixes #*
- [x] Explain the changes made
- [ ] Sign the commit using for [DCO](https://github.com/apps/dco) 
- [ ] Rebase branch to HEAD

---------

Signed-off-by: Andres Morey <andresmarcel@gmail.com>