🐋 Add option to dev environment to compile Rust code incrementally on host machine (#288)
## Summary

Added an optional flag to tilt up that allows use of local builds for
rgkl as opposed to containerized builds, increasing initial compile time
by about 2x (60s -> 30s) and subsequent compile-times by 30x (60s ->
~2s). This will hopefully provide a solid base for growth in and
contribution to the Rust sub-component(s) of the project.

## Changes

- Modify dev and production cluster-agent docker images to copy
executable binary and shared libraries so `rgkl` can be called without
explicit call to linker
- ⁠Use static compilation (musl) in local builds. This allows us to
standardize on the executable itself.
- ⁠⁠Call executable directly instead of using a loader in containerized
builds (dev and prod). This change was required as loaders do not use
musl and we have a single entrypoint (⁠ logrecords.go ⁠) in the system
as a whole.
- ⁠Provide documentation on how to use this feature.

---------

Signed-off-by: Andres Morey <andresmarcel@gmail.com>
Co-authored-by: Sinan C. Gulan <57106474+sinancang@users.noreply.github.com>